+++ source/jormungandr/jormungandr/pt_planners/common.py
@@ -42,8 +42,7 @@ from abc import abstractmethod, ABCMeta
 
 from jormungandr import app
 from jormungandr.exceptions import DeadSocketException
-from navitiacommon import response_pb2, type_pb2, request_pb2
-from jormungandr import utils
+from navitiacommon import response_pb2
 
 
 class ZmqSocket(six.with_metaclass(ABCMeta, object)):
@@ -115,105 +114,3 @@ class ZmqSocket(six.with_metaclass(ABCMeta, object)):
             return self.breaker.call(self._send_and_receive, *args, **kwargs)
         except pybreaker.CircuitBreakerError:
             raise DeadSocketException(self.name, self.zmq_socket)
-
-
-def create_journeys_request(origins, destinations, datetime, clockwise, journey_parameters, bike_in_pt):
-    req = request_pb2.Request()
-    req.requested_api = type_pb2.pt_planner
-    for stop_point_id, access_duration in origins.items():
-        location = req.journeys.origin.add()
-        location.place = stop_point_id
-        location.access_duration = access_duration
-
-    for stop_point_id, access_duration in destinations.items():
-        location = req.journeys.destination.add()
-        location.place = stop_point_id
-        location.access_duration = access_duration
-
-    req.journeys.night_bus_filter_max_factor = journey_parameters.night_bus_filter_max_factor
-    req.journeys.night_bus_filter_base_factor = journey_parameters.night_bus_filter_base_factor
-
-    req.journeys.datetimes.append(datetime)
-    req.journeys.clockwise = clockwise
-    req.journeys.realtime_level = utils.realtime_level_to_pbf(journey_parameters.realtime_level)
-    req.journeys.max_duration = journey_parameters.max_duration
-    req.journeys.max_transfers = journey_parameters.max_transfers
-    req.journeys.wheelchair = journey_parameters.wheelchair
-    if journey_parameters.max_extra_second_pass:
-        req.journeys.max_extra_second_pass = journey_parameters.max_extra_second_pass
-
-    for uri in journey_parameters.forbidden_uris:
-        req.journeys.forbidden_uris.append(uri)
-
-    for id in journey_parameters.allowed_id:
-        req.journeys.allowed_id.append(id)
-
-    if journey_parameters.direct_path_duration is not None:
-        req.journeys.direct_path_duration = journey_parameters.direct_path_duration
-
-    req.journeys.bike_in_pt = bike_in_pt
-
-    if journey_parameters.min_nb_journeys:
-        req.journeys.min_nb_journeys = journey_parameters.min_nb_journeys
-
-    if journey_parameters.timeframe:
-        req.journeys.timeframe_duration = int(journey_parameters.timeframe)
-
-    if journey_parameters.depth:
-        req.journeys.depth = journey_parameters.depth
-
-    if journey_parameters.isochrone_center:
-        req.journeys.isochrone_center.place = journey_parameters.isochrone_center
-        req.journeys.isochrone_center.access_duration = 0
-        req.requested_api = type_pb2.ISOCHRONE
-
-    if journey_parameters.sn_params:
-        sn_params_request = req.journeys.streetnetwork_params
-        sn_params_request.origin_mode = journey_parameters.sn_params.origin_mode
-        sn_params_request.destination_mode = journey_parameters.sn_params.destination_mode
-        sn_params_request.walking_speed = journey_parameters.sn_params.walking_speed
-        sn_params_request.bike_speed = journey_parameters.sn_params.bike_speed
-        sn_params_request.car_speed = journey_parameters.sn_params.car_speed
-        sn_params_request.bss_speed = journey_parameters.sn_params.bss_speed
-        sn_params_request.car_no_park_speed = journey_parameters.sn_params.car_no_park_speed
-
-    if journey_parameters.current_datetime:
-        req._current_datetime = journey_parameters.current_datetime
-
-    if journey_parameters.walking_transfer_penalty:
-        req.journeys.walking_transfer_penalty = journey_parameters.walking_transfer_penalty
-
-    if journey_parameters.arrival_transfer_penalty:
-        req.journeys.arrival_transfer_penalty = journey_parameters.arrival_transfer_penalty
-
-    return req
-
-
-def create_graphical_isochrones_request(
-    origins, destinations, datetime, clockwise, graphical_isochrones_parameters, bike_in_pt
-):
-    req = create_journeys_request(
-        origins,
-        destinations,
-        datetime,
-        clockwise,
-        graphical_isochrones_parameters.journeys_parameters,
-        bike_in_pt,
-    )
-    req.requested_api = type_pb2.graphical_isochrone
-    req.journeys.max_duration = graphical_isochrones_parameters.journeys_parameters.max_duration
-    if graphical_isochrones_parameters.boundary_duration:
-        for duration in sorted(graphical_isochrones_parameters.boundary_duration, key=int, reverse=True):
-            if graphical_isochrones_parameters.min_duration < duration < req.journeys.max_duration:
-                req.isochrone.boundary_duration.append(duration)
-    req.isochrone.boundary_duration.insert(0, req.journeys.max_duration)
-    req.isochrone.boundary_duration.append(graphical_isochrones_parameters.min_duration)
-
-    # We are consistent with new_default there.
-    if req.journeys.origin:
-        req.journeys.clockwise = True
-    else:
-        req.journeys.clockwise = False
-
-    req.isochrone.journeys_request.CopyFrom(req.journeys)
-    return req
