+++ source/jormungandr/jormungandr/pt_planners/kraken.py
@@ -28,11 +28,8 @@
 # www.navitia.io
 from __future__ import absolute_import, print_function, unicode_literals, division
 
-from jormungandr.pt_planners.common import (
-    ZmqSocket,
-    create_journeys_request,
-    create_graphical_isochrones_request,
-)
+from jormungandr.pt_planners.common import ZmqSocket
+from jormungandr import utils
 from .pt_planner import AbstractPtPlanner
 
 
@@ -48,13 +45,15 @@ class Kraken(ZmqSocket, AbstractPtPlanner):
         raise NotImplementedError("Too bad, you cannot ask kraken for places nearby :)")
 
     def journeys(self, origins, destinations, datetime, clockwise, journey_parameters, bike_in_pt, request_id):
-        req = create_journeys_request(origins, destinations, datetime, clockwise, journey_parameters, bike_in_pt)
+        req = utils.create_journeys_request(
+            origins, destinations, datetime, clockwise, journey_parameters, bike_in_pt
+        )
         return self.send_and_receive(req, request_id=request_id)
 
     def graphical_isochrones(
         self, origins, destinations, datetime, clockwise, graphical_isochrones_parameters, bike_in_pt
     ):
-        req = create_graphical_isochrones_request(
+        req = utils.create_graphical_isochrones_request(
             origins, destinations, datetime, clockwise, graphical_isochrones_parameters, bike_in_pt
         )
         return self.send_and_receive(req)
